+++ backend/promotion-service/src/main/java/id/payu/promotion/domain/model/PromoCode.java
@@ -113,7 +113,7 @@ public class PromoCode {
 
         if (discountType == DiscountType.PERCENTAGE) {
             discount = amount.multiply(discountValue)
-                    .divide(BigDecimal.valueOf(100), 2, RoundingMode.HALF_EVEN);
+                    .divide(BigDecimal.valueOf(100), 4, RoundingMode.HALF_EVEN);
         } else {
             discount = discountValue;
         }